resolve short links
